change js and create new main required files
