+++ js/main-page/main-page.js
@@ -1,0 +1,3 @@
+require('./validator.js');
+require('./mask.js');
+require('./mobile-menu.js');
