+++ js/script.js
@@ -1,5 +1,2 @@
-require('./main-page/validator.js');
-require('./main-page/mask.js');
-require('./main-page/mobile-menu.js');
-require('./slider/show-images.js');
-require('./slider/next-prev.js');
+require('./main-page/main-page.js');
+require('./slider/slider.js');
