+++ js/slider/slider.js
@@ -1,2 +1,3 @@
-// require('./slider-images');
-// require('./slider-pagination');
+require('./show-images.js');
+require('./next-prev.js');
+require('./pagination.js');
